quick refactor :)
